requirements.txt
requests
python-dotenv
psycopg2-binary
sqlalchemy-redshift
redshift-connector
pandas
pytz